Added RLAgent.
Added RLAgent, a simple controllable agent that gets waypoints and goes
where you tell it to.
Added in the diffuse lighting color changes from Project 1. This touches
a number of files.
Added the capability to change the waypoint list debug drawing color, so
the mouse can have blue lines and the cat can have red lines. This is in
Movement.cpp.
Modified World.cpp to create two instances of RLAgent, "Cat0" and
"Mouse0".
Also modified World.cpp to add the capability of generating more cats
and mice if we so choose.
Added the messages necessary for the RLAgent to msgnames.h.
Added functions for waypoint manipulation to the Movement class.